geekshop/8+++

+++ geekshop/adminapp/views.py
@@ -1,9 +1,11 @@
 from django.shortcuts import render, HttpResponseRedirect
 from authapp.models import User
 from mainapp.models import ProductCategory
-from adminapp.forms import UserAdminRegisterForm, UserAdminProfileForm, CategoryAdminRegisterForm
+from adminapp.forms import UserAdminRegisterForm, UserAdminProfileForm, CategoryAdminRegisterForm, \
+    CategoryAdminUpdateForm
 from django.urls import reverse
 from django.contrib.auth.decorators import user_passes_test
+from django.contrib import messages
 
 
 # Create your views here.
@@ -30,11 +32,11 @@ def admin_users_create(request):
         else:
             return HttpResponseRedirect(reverse('admin_staff:admin_category_create'))
     form = UserAdminRegisterForm()
-    contex = {
+    context = {
         'form': form,
     }
 
-    return render(request, 'adminapp/admin-users-create.html', contex)
+    return render(request, 'adminapp/admin-users-create.html', context)
 
 
 @user_passes_test(lambda user: user.is_superuser)
@@ -67,6 +69,8 @@ def admin_users_remove(request, user_id=None):
     return HttpResponseRedirect(reverse('admin_staff:admin_users'))
 
 
+# ==============================categories=============================================================== #
+
 def admin_category(request):
     categories = ProductCategory.objects.all()
     context = {
@@ -82,33 +86,39 @@ def admin_category_create(request):
             form.save()
             return HttpResponseRedirect(reverse('admin_staff:admin_category'))
         else:
+            messages.error('Что-то не так!!!')
             return HttpResponseRedirect(reverse('admin_staff:admin_category'))
     form = CategoryAdminRegisterForm()
-    contex = {
+    context = {
         'form': form,
     }
-
-    return render(request, 'adminapp/admin-category-create.html', contex)
+    return render(request, 'adminapp/admin-category-create.html', context)
 
 
 def admin_category_update(request, categ_id):
     category = ProductCategory.objects.get(id=categ_id)
     if request.method == 'POST':
-        form = CategoryAdminRegisterForm(data=request.POST)
+        form = CategoryAdminUpdateForm(data=request.POST)
         if form.is_valid():
             form.save()
+
             return HttpResponseRedirect(reverse('admin_staff:admin_category'))
         else:
             return HttpResponseRedirect(reverse('admin_staff:admin_category'))
-    form = CategoryAdminRegisterForm()
-    contex = {
+    form = CategoryAdminUpdateForm()
+    context = {
         'form': form,
         'category': category,
     }
-    return render(request, 'adminapp/admin-category-update-delete.html', contex)
+    return render(request, 'adminapp/admin-category-update-delete.html', context)
+
 
 def admin_category_delete(request, categ_id):
+    print(f'--------------------------------------------------{categ_id}')
     category = ProductCategory.objects.get(id=categ_id)
-    category.delete()
+    print(f'-------------------------------------------------->{category}')
+    if category:
+        category.delete()
+        return HttpResponseRedirect(reverse('admin_staff:admin_category'))
 
     return render(request, 'adminapp/admin-category-update-delete.html')
